fix ATR-12192 (TrigT1CaloSim-00-02-29-14)
	* CPMTowerMaker: fix ATR-12192
	* Tag as TrigT1CaloSim-00-02-29-14

2015-09-08 Veit Scharf <veit.scharf@kip.uni-heidelberg.de>
	* TrigT1CaloSim_ReadTT.py: fix ATR-12162
	* Tag as TrigT1CaloSim-00-02-29-13

2015-09-02 Alan Watson <Alan.Watson@CERN.CH>
        * EnergyCMX: Protection against ATR-11137 bug (as in 00-03-01)
        * Tag as TrigT1CaloSim-00-02-29-12

2015-09-02 Alan Watson <Alan.Watson@CERN.CH>
        * JetCMX: incorporate changes from TrigT1CaloSim-00-03-01
	* Tag as TrigT1CaloSim-00-02-29-11

+++ Trigger/TrigT1/TrigT1CaloSim/src/CPMTowerMaker.cxx
@@ -107,8 +107,8 @@ StatusCode CPMTowerMaker::execute( )
   if (msgLvl(MSG::DEBUG)) {
     ATH_MSG_DEBUG ( "Formed " << vectorOfCPMTs->size() << " CPM Towers " ) ;
     for (const CPMTower* cpmt : *vectorOfCPMTs)
-      ATH_MSG_DEBUG ( "CPMT has coords (" << cpmt->eta() <<", "<< cpmt->phi() << ") and energies : "
-                      << cpmt->emEnergy() <<", "<< cpmt->hadEnergy() <<" (Em,Had)" );
+      ATH_MSG_DEBUG ( "CPMT has coords (" << cpmt->eta() << ", "<< cpmt->phi() << ") and energies : "
+                      << static_cast<unsigned int>(cpmt->emEnergy()) << ", "<< static_cast<unsigned int>(cpmt->hadEnergy()) <<" (Em,Had)" );
 
   }
